wheelie of fortune started
