+++ client/package.json
@@ -9,6 +9,7 @@
     "@testing-library/user-event": "^12.8.3",
     "bcryptjs": "^2.4.3",
     "date-fns": "^2.29.3",
+    "example": "^0.0.0",
     "formik": "^2.2.9",
     "react": "^17.0.2",
     "react-big-calendar": "^1.6.9",
@@ -18,6 +19,7 @@
     "react-markdown": "^6.0.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^5.0.1",
+    "react-wheel-of-prizes": "^1.0.9",
     "styled-components": "^5.3.0",
     "web-vitals": "^1.1.2",
     "yup": "^1.0.2"
