+++ client/src/App.js
@@ -76,7 +76,7 @@ function App() {
           <ChefsContainer chefs={chefs}/>
         </Route>
         <Route path="/calendar" >
-          <CalendarPage />
+          <CalendarPage recipes={recipes}/>
         </Route>
         <Route path='/reviews'>
           <ReviewsContainer reviews={reviews} addReview={addReview}/>
